# docs/swagger.yaml
definitions:
...
  model.CreateLessonInput:
    properties:
      group_id:
        type: string
      lecture_room:
        type: string
      lesson_type:
        type: string
      subject_id:
        type: integer
      teacher_id:
        type: integer
      time_end:
        type: string
      time_start:
        type: string
    type: object
  model.CreatePersonInput:
    properties:
      birth_date:
        type: string
      name:
        type: string
      phone:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
  model.Lesson:
    properties:
      address:
        type: string
      birth_date:
        type: string
      email:
        type: string
      employee_id:
        type: integer
      faculty_id:
        type: string
      full_name:
        type: string
      group_id:
        type: string
      hire_date:
        type: string
      last_visit_at:
        type: string
      lecture_room:
        type: string
      lesson_id:
        type: integer
      lesson_type:
        type: string
      name:
        type: string
      notes:
        type: string
      password:
        type: string
      person_id:
        type: integer
      phone:
        type: string
      photo:
        items:
          type: integer
        type: array
      registered_at:
        type: string
      salary:
        type: number
      specialty_id:
        type: integer
      subject_id:
        type: integer
      surname:
        type: string
      teacher_id:
        type: integer
      time_end:
        type: string
      time_start:
        type: string
      title:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  model.Person:
    properties:
      address:
        type: string
      birth_date:
        type: string
      email:
        type: string
      last_visit_at:
        type: string
      name:
        type: string
      notes:
        type: string
      password:
        type: string
      person_id:
        type: integer
      phone:
        type: string
      photo:
        items:
          type: integer
        type: array
      registered_at:
        type: string
      surname:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  model.UpdatePersonInput:
    properties:
      address:
        type: string
      birth_date:
        type: string
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      photo:
        items:
          type: integer
        type: array
      surname:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for test.
  title: University System API
  version: "2.0"
paths:
...
  /admin/hub/lessons:
    get:
      consumes:
      - application/json
      description: get lessons
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            items:
              $ref: '#/definitions/model.Lesson'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get lessons
      tags:
      - admin-lessons
    post:
      consumes:
      - application/json
      description: create lesson
      parameters:
      - description: Data for creating lesson
        in: body
        name: lesson
        required: true
        schema:
          $ref: '#/definitions/model.CreateLessonInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create lesson
      tags:
      - admin-lessons
  /admin/hub/lessons/{id}:
    delete:
      consumes:
      - application/json
      description: delete lesson
      parameters:
      - description: ID for deleting lesson
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete lesson
      tags:
      - admin-lessons
    get:
      consumes:
      - application/json
      description: get lesson by id
      parameters:
      - description: ID for getting lesson
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            $ref: '#/definitions/model.Lesson'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get lesson
      tags:
      - admin-lessons
  /admin/hub/people:
    get:
      consumes:
      - application/json
      description: get people
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            items:
              $ref: '#/definitions/model.Person'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get people
      tags:
      - admin-people
    post:
      consumes:
      - application/json
      description: create person
      parameters:
      - description: Data for creating person
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/model.CreatePersonInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create person
      tags:
      - admin-people
  /admin/hub/people/{id}:
    delete:
      consumes:
      - application/json
      description: delete person
      parameters:
      - description: ID for deleting person
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete person
      tags:
      - admin-people
    get:
      consumes:
      - application/json
      description: get person by id
      parameters:
      - description: ID for getting person
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            $ref: '#/definitions/model.Person'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get person
      tags:
      - admin-people
    patch:
      consumes:
      - application/json
      description: update person
      parameters:
      - description: ID for updating person
        in: path
        name: id
        required: true
        type: integer
      - description: New information for update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePersonInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update person
      tags:
      - admin-people
...
swagger: "2.0"